Use generated ConstraintSetModel.
